backend complete

+++ backend/controllers/goalController.js
@@ -1,11 +1,11 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
-
+const User = require('../models/userModel');
 // $desc     Get all goals
 // $route    GET /api/goals
 // $access   Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user._id  });
     res.status(200).json(goals);
 })
 
@@ -20,7 +20,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add a text field');
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user._id,
     })
 
     res.status(200).json(goal);
@@ -37,6 +38,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found');
     }
 
+    const user = await User.findById(req.user._id);
+
+    // Check for user
+    if(!user){
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Make sure the logged in user matches the goal user.
+    if(goal.user.toString() !== req.user._id.toString()){
+        res.status(401);
+        throw new Error('User not authorized.');
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
     res.status(200).json(updatedGoal);
@@ -51,6 +66,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400);
         throw new Error('Goal not found');
+    }
+
+    const user = await User.findById(req.user._id);
+
+    // Check for user
+    if(!user){
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Make sure the logged in user matches the goal user.
+    if(goal.user.toString() !== req.user._id.toString()){
+        res.status(401);
+        throw new Error('User not authorized.');
     }
     
     await Goal.findByIdAndDelete(req.params.id);
